Added LCD update API
This change rewrites how LCD state is stored. LCD state is now kept in a
separate structure, and sent to a callback with a monotonic timestamp
since system boot, excepting time spent in the spin function.

The callback will only be triggered when the display state has changed,
though it is advisable to at least filter events so that only the most
recent which has occurred in a given time frame is logged and rendered.

The display also provides a rendering function which can be called on a
const state block, which will produce one 1024x1024 RGB bitmap for the
caller to do with what they want.